cherry-picking all bug fixes from cloud_6_30 branch to master (#13999)
* RFC bug fixes (#13787)


(cherry picked from commit c901bce8aa0ec1df98670030bfa389335e2f27fc)

* making default flow scale in throttle to 8M * 2 (#13788)


(cherry picked from commit 280304afa15c26b7f480177b4b74c8a683a314a4)

* RFC fixes - part 2 (#13801)


(cherry picked from commit 644e020055db701bc57566027027dd5b86148813)

* die temparature log fix (#13821)


(cherry picked from commit 533963aeb427551d5de9fbedace0d2c56bd7976b)

* fix for 32 bit route result packing issue (#13875)


(cherry picked from commit e67e4de4f614ebb96af7e17e220c712cc93cbbf1)

+++ nic/apollo/api/impl/artemis/lpm/lpm_ipv4_peer_route.cc
@@ -82,44 +82,37 @@ lpm_ipv4_peer_route_add_key_to_last_stage (uint8_t *bytes, uint32_t idx,
         case 0:
             table->action_u.txlpm1_match1_32b_retrieve.key0 =
                     lpm_inode->ipaddr.addr.v4_addr;
-            table->action_u.txlpm1_match1_32b_retrieve.data0 =
-                    (uint16_t)lpm_inode->data;
+            table->action_u.txlpm1_match1_32b_retrieve.data0 = lpm_inode->data;
             break;
         case 1:
             table->action_u.txlpm1_match1_32b_retrieve.key1 =
                     lpm_inode->ipaddr.addr.v4_addr;
-            table->action_u.txlpm1_match1_32b_retrieve.data1 =
-                    (uint16_t)lpm_inode->data;
+            table->action_u.txlpm1_match1_32b_retrieve.data1 = lpm_inode->data;
             break;
         case 2:
             table->action_u.txlpm1_match1_32b_retrieve.key2 =
                     lpm_inode->ipaddr.addr.v4_addr;
-            table->action_u.txlpm1_match1_32b_retrieve.data2 =
-                    (uint16_t)lpm_inode->data;
+            table->action_u.txlpm1_match1_32b_retrieve.data2 = lpm_inode->data;
             break;
         case 3:
             table->action_u.txlpm1_match1_32b_retrieve.key3 =
                     lpm_inode->ipaddr.addr.v4_addr;
-            table->action_u.txlpm1_match1_32b_retrieve.data3 =
-                    (uint16_t)lpm_inode->data;
+            table->action_u.txlpm1_match1_32b_retrieve.data3 = lpm_inode->data;
             break;
         case 4:
             table->action_u.txlpm1_match1_32b_retrieve.key4 =
                     lpm_inode->ipaddr.addr.v4_addr;
-            table->action_u.txlpm1_match1_32b_retrieve.data4 =
-                    (uint16_t)lpm_inode->data;
+            table->action_u.txlpm1_match1_32b_retrieve.data4 = lpm_inode->data;
             break;
         case 5:
             table->action_u.txlpm1_match1_32b_retrieve.key5 =
                     lpm_inode->ipaddr.addr.v4_addr;
-            table->action_u.txlpm1_match1_32b_retrieve.data5 =
-                    (uint16_t)lpm_inode->data;
+            table->action_u.txlpm1_match1_32b_retrieve.data5 = lpm_inode->data;
             break;
         case 6:
             table->action_u.txlpm1_match1_32b_retrieve.key6 =
                     lpm_inode->ipaddr.addr.v4_addr;
-            table->action_u.txlpm1_match1_32b_retrieve.data6 =
-                    (uint16_t)lpm_inode->data;
+            table->action_u.txlpm1_match1_32b_retrieve.data6 = lpm_inode->data;
             break;
         default:
             break;
@@ -132,8 +125,7 @@ sdk_ret_t
 lpm_ipv4_peer_route_set_default_data (uint8_t *bytes, uint32_t default_data)
 {
     auto table = (txlpm1_actiondata_t *) bytes;
-    table->action_u.txlpm1_match1_32b_retrieve.data_ =
-            (uint16_t)default_data;
+    table->action_u.txlpm1_match1_32b_retrieve.data_ = default_data;
     return SDK_RET_OK;
 }
